requirements.txt
pandas==2.0.0
google-cloud-bigtable==2.16.0
google-auth==2.17.3 